Foro funcional

+++ src/main/java/com/example/accessingdatamysql/comment/CommentRepository.java
@@ -2,6 +2,6 @@ package com.example.accessingdatamysql.comment;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface CommentRepository extends CrudRepository<Comment, Integer> {
+public interface CommentRepository extends CrudRepository<Comment, Long> {
 
 }
