+++ src/main/java/com/example/accessingdatamysql/forum/ForumRepository.java
@@ -2,6 +2,6 @@ package com.example.accessingdatamysql.forum;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ForumRepository extends CrudRepository<Forum, Integer> {
+public interface ForumRepository extends CrudRepository<Forum, Long> {
 
 }
